+++ docs/Changelog.md
@@ -21,7 +21,8 @@ Nussknacker versions
 * [#2441](https://github.com/TouK/nussknacker/pull/2441) avro sink supports defaults of primitive avro types
 * [#2498](https://github.com/TouK/nussknacker/pull/2498), [#2499](https://github.com/TouK/nussknacker/pull/2499), [#2503](https://github.com/TouK/nussknacker/pull/2503), [#2539](https://github.com/TouK/nussknacker/pull/2539) EspExceptionHandler is removed from ProcessConfigCreator.
   Flink engine uses now fixed exception handler: FlinkExceptionHandler. All deprecated FlinkEspExceptionHandler implementations are removed.
-     
+* [#2543](https://github.com/TouK/nussknacker/pull/2543) Eager parameters can have helpers injected.
+
 1.1.0 (Not released yet)
 ------------------------
 * [#2176](https://github.com/TouK/nussknacker/pull/2176) Allow to enrich periodic scenario config on initial schedule and each deployment.